More incremental 9p changes.
The function handling is more-or-less working, but it appears that conv_S2M() is lacking.

It's entirely possible that I'm just not smart enough to grok the original implementation, but
I've been converting it as I go to something I can understand better.

I should probably check in my test filesystem as an example. Maybe once it's further along.